+++ src/components/feedback-home/colors.js
@@ -1,0 +1,9 @@
+export const COLORS = {
+  bg: "#F7F6F3",
+  cardBg: "rgba(255,255,255,0.7)",
+  cardBorder: "#EAE6DF",
+  text: "#2F2A25",
+  subtext: "#7C736A",
+  accent: "#0a2259",
+  shadow: "0 12px 35px rgba(0,0,0,0.08)",
+};
